fix owner permission

+++ controllers/contacts.js
@@ -19,13 +19,11 @@ const getAll = async (req, res) => {
 const getById = async (req, res) => {
   const { _id: owner } = req.user;
   const { contactId } = req.params;
-  const result = await Contact.findById(contactId);
+  const result = await Contact.findById({ contactId, owner });
   if (!result) {
     throw HttpError(404, "Not found");
   }
-  if (owner !== result._id) {
-    throw HttpError(403, "Forbidden, this contact does not belong to you ");
-  }
+
   res.status(200).json(result);
 };
 
@@ -39,11 +37,8 @@ const add = async (req, res) => {
 const deleteById = async (req, res) => {
   const { _id: owner } = req.user;
   const { contactId } = req.params;
-  const itemToDell = await Contact.findById(contactId);
-  if (itemToDell._id !== owner) {
-    throw HttpError(403, "Forbidden, this contact does not belong to you");
-  }
-  const result = await Contact.findByIdAndDelete(contactId);
+
+  const result = await Contact.findByIdAndDelete({ contactId, owner });
   if (!result) {
     throw HttpError(404, "Not found");
   }
@@ -53,13 +48,14 @@ const deleteById = async (req, res) => {
 const updateById = async (req, res) => {
   const { _id: owner } = req.user;
   const { contactId } = req.params;
-  const itemToUpd = await Contact.findById(contactId);
-  if (itemToUpd._id !== owner) {
-    throw HttpError(403, "Forbidden, this contact does not belong to you");
-  }
-  const result = await Contact.findByIdAndUpdate(contactId, req.body, {
-    new: true,
-  });
+
+  const result = await Contact.findByIdAndUpdate(
+    { contactId, owner },
+    req.body,
+    {
+      new: true,
+    }
+  );
 
   if (!result) {
     throw HttpError(404, "Not found");
@@ -70,13 +66,13 @@ const updateById = async (req, res) => {
 const updateStatusContact = async (req, res) => {
   const { _id: owner } = req.user;
   const { contactId } = req.params;
-  const itemToUpd = await Contact.findById(contactId);
-  if (itemToUpd._id !== owner) {
-    throw HttpError(403, "Forbidden, this contact does not belong to you");
-  }
-  const result = await Contact.findByIdAndUpdate(contactId, req.body, {
-    new: true,
-  });
+  const result = await Contact.findByIdAndUpdate(
+    { contactId, owner },
+    req.body,
+    {
+      new: true,
+    }
+  );
 
   if (!result) {
     throw HttpError(404, "Not found");
